fix(extension): isolate scanner from auth surfaces
