+++ apps/vela-ext/entrypoints/content.ts
@@ -1,5 +1,3 @@
-import { readCognitoSessionFromStorage } from './utils/webappSession';
-
 // Export for unit testing
 export function scanJapaneseSentences(): string[] {
   const japaneseRe = /[\u3040-\u9FAF\uFF66-\uFF9F]/;
@@ -315,39 +313,34 @@ function buildOverlay(sentences: string[]): ShadowRoot {
 // Content script entrypoint
 export default defineContentScript({
   matches: ['*://*/*'],
+  excludeMatches: [
+    'https://vela.cwchanap.dev/*',
+    'http://localhost:9000/*',
+    'http://127.0.0.1:9000/*',
+    'https://*.amazoncognito.com/*',
+    'https://*.google.com/*',
+  ],
   main() {
-    browser.runtime.onMessage.addListener(
-      (message: unknown, _sender: unknown, sendResponse: (_response?: unknown) => void) => {
-        if (
-          typeof message !== 'object' ||
-          message === null ||
-          !('type' in message) ||
-          (message.type !== 'SCAN_PAGE' && message.type !== 'GET_VELA_WEBAPP_SESSION')
-        )
-          return;
+    browser.runtime.onMessage.addListener((message: unknown) => {
+      if (typeof message !== 'object' || message === null || !('type' in message)) {
+        return;
+      }
+      if (message.type !== 'SCAN_PAGE') {
+        return;
+      }
 
-        if (message.type === 'GET_VELA_WEBAPP_SESSION') {
-          try {
-            sendResponse(readCognitoSessionFromStorage(window.localStorage));
-          } catch {
-            sendResponse(null);
-          }
-          return;
-        }
+      // Remove any existing overlay before creating a new one
+      document.getElementById('vela-ext-overlay-host')?.remove();
 
-        // Remove any existing overlay before creating a new one
-        document.getElementById('vela-ext-overlay-host')?.remove();
+      const sentences = scanJapaneseSentences();
+      if (sentences.length === 0) {
+        // Content scripts can't use browser.notifications — ask the background
+        // script to show the "no sentences found" notification.
+        browser.runtime.sendMessage({ type: 'NO_JAPANESE_FOUND' }).catch(() => {});
+        return;
+      }
 
-        const sentences = scanJapaneseSentences();
-        if (sentences.length === 0) {
-          // Content scripts can't use browser.notifications — ask the background
-          // script to show the "no sentences found" notification.
-          browser.runtime.sendMessage({ type: 'NO_JAPANESE_FOUND' }).catch(() => {});
-          return;
-        }
-
-        buildOverlay(sentences);
-      },
-    );
+      buildOverlay(sentences);
+    });
   },
 });
