+++ apps/vela-ext/types/globals.d.ts
@@ -1,7 +1,11 @@
 // Type declarations for WXT globals
 declare global {
   function defineBackground(_callback: () => void): void;
-  function defineContentScript(_config: { matches: string[]; main(): void }): void;
+  function defineContentScript(_config: {
+    matches: string[];
+    excludeMatches?: string[];
+    main(): void;
+  }): void;
 }
 
 export {};
